app/pyproject.toml
[tool.poetry]
name = "tfs"
version = "0.1.0"
description = "Temporary File Storage"
authors = ["saasan"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.71.0"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
aiomysql = "^0.0.22"
SQLAlchemy = "^1.4.29"
ulid-py = "^1.1.0"
python-multipart = "^0.0.5"
aiofiles = "^0.8.0"
certifi = "^2021.10.8"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.16.0"
aiosqlite = "^0.17.0"
httpx = "^0.21.3"
freezegun = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
